+++ platform/jvm/capture-plugin/src/main/kotlin/io/bitdrift/capture/extension/InstrumentationExtension.kt
@@ -47,7 +47,7 @@ open class InstrumentationExtension
             objects
                 .property(
                     OkHttpInstrumentationType::class.java,
-                ).convention(OkHttpInstrumentationType.OVERWRITE)
+                ).convention(OkHttpInstrumentationType.PROXY)
 
         enum class OkHttpInstrumentationType {
             PROXY,